Refactor startResizeAnimation in PipTransition
In PiP2, we want to allow for the animator to manipulate
startTx (and apply startTx) and finishTx and update PipTransitionState
to let PipTransition finishTransition() while continuing playing
the animation after.

See go/pip2-transitions PiP Double Tap CUJ > Approach #2

This is done by updating the state and sending an extra bundle
to a subscriber handling animator scheduling. The bundle contains
startTx, finishTx, and destination bounds
(in case the animator is somehow already finished).

This should reduce the latency of PiP interactions such as
pinching and double tapping as describe in the design doc.

Also added SWIPING_TO_PIP to stateToString(), otherwise we can
get IllegalStateException when swiping up.

Bug: 332770512
Test: N/A until connected to PipTouchHandler e.g.
Test: atest WMShellUnitTests:PipTransitionStateTest
Change-Id: I1dc4f709f76bdfa50584153c41f2af224b03ba3b