Update time-limited FGS logic to handle parallel services.
If there are multiple foreground services running of the same type
from the same uid, update the total runtime immediately when another
parallel service of the same type starts. Also keep a counter of how
many services are running in parallel to reset the tracking properly.

Also remove redundant calls to the stop-fgs-tracking method which could
affect the new tracking logic.

Bug: 330399444
Test: atest CtsFgsTimeoutTestCases
Change-Id: I9681b5187406eb755016ee6b86cd4162f54a26c1