Remove PM#getPackageInfo call in restoreAtInstall flow
A previous change (ag/26748034) added this call for BMM logging. getPackageInfo is a relatively heavy binder call, which triggered performance regression alerts (b/342082681 and b/342273383). The additional getPackageInfo call resulted in 5-10% increase in latency for package commit operation.

For the purposes of BMM logging in restoreAtInstall flow, we just need the name of the package. Fields like version code are already logged in subsequent BMM logging during restore.

Bug: 342082681
Test: Verified that logging is correct

Change-Id: I8025789b2c696eac1b7738a1ae332c67aea8ce5f