Use non-protected screenshot buffer to calculate animation color
The screenshot for animation needs to be "setAllowProtected(true)",
otherwise the protected area will be blank. But it is unable to
be used to extract color because protected content is not readable.

To avoid the background rotation animation always showing black
color in this case (it looks flickering if the main color of screen
is light), use the source surface with excluding protected content
to calculate the color.

Bug: 342984506
Test: Enable auto rotation, home rotation.
      Set wallpaper to white color.
      Play DRM protected video in PiP mode.
      Rotate the device.
      The animation background should be white.
Change-Id: I346c9683e2c58891103031bf9adf662a7044132c