Resize animation after pinching PiP
Implement resize animator after pinch-release
in PiP2. Make sure that PipTransitionState is
properly set to CHANGED_PIP_BOUNDS only when
the new buffer has arrived and Shell received a signal for it
as a part of the config-at-end transition.

The latter makes sure that PiP touch interactions are not enabled
until the transition is fully complete, and all the relevant state
is up-to-date and ready for the next interaction.

Bug: 346586058
Test: swipe-up enter PiP2 and pinch-resize PiP window

Change-Id: I74607950f20d17953d08b16e3076f0407d8e1aa0