Rename IMMS#mPackageMonitorHandler to mIoHandler
This is a pure refactoring CL to rename

  InputMethodManagerService#mPackageMonitorHandler

to

  InputMethodManagerService#mIoHandler

to make it clear that we are going to use it not only from

  InputMethodManagerService.MyPackageMonitor

but also from other controllers where we want to perform I/O
operations without holding ImfLock.class.

This is a pure renaming. There must be no observable side effect.

Bug: 343601565
Test: presubmit
Flag: EXEMPT refactor
Change-Id: I3a0e1f27d3dc9c7db68693f1c4a50dfc5c83358e