[flexiglass] Bouncer scene cannot be occluded
When an occluding activity is shown on top of the lockscreen (for
example the camera app through double-tap power button) and then
something shows the bouncer, the bouncer scene should be shown on top of
the occluding activity.

To make that happens, we make the bouncer scene be considered as a scene
that can't be occluded.

Fix: 349438272
Test: added unit tests expanded
Test: manually verified that tapping on the lock icon in the camera app
correctly shows the bouncer
Flag: com.android.systemui.scene_container

Change-Id: Ia1f68dc19e8a7a20f137ba76229ebdd1e08b35a6