Fix NPE in ApplicationExitInfoTest
ApplicationExitInfoTest discards ProcessList.sKillHandler so that a
new one can be created.  However, if a message is in flight on the
handler when it is discarded, an NPE will result.

KillHandler.handleMessage() no longer refers to the global sKillHandler,
which may be null (the source of the NPE).  handleMessage() instead
refers to this, which can never be null.  This has no impact to normal
operation because sKillHandler is never altered.

Test: atest
 * FrameworksMockingServicesTests:com.android.server.am
 * FrameworksServicesTests:com.android.server.am

Flag: EXEMPT bugfix
Bug: 349349773
Change-Id: Iaa333e314228c12fbfd794cef1a2ebf104e939fb