Add unlock reason to alternate bouncer when starting activity
Pass a custom message parameter to the APIs used by the glanceable hub
to start an activity that dismisses the keyguard. This message will
serve as the unlock reason in the alternate bouncer.
On the communal side, support the two use cases for displaying the
unlock reason: 1) when starting the EditWidgetsActivity;
2) tapping on a widget that requires authentication.

Bug: b/325110448
Flag: com.android.systemui.communal_hub
Test: atest LegacyActivityStarterInternalImplTest
Test: atest WidgetInteractionHandlerTest
Test: manually with device_entry_udfps_refactor flag off
Change-Id: Ib89c0b1a6b0cfef0e7f7b5f0b7770f58f4ea689e