Cache user serial numbers
This CL adds a cache for user serial numbers in UserManager. This is done to improve performance, as the UserManagerService can be slow to respond to requests for user serial numbers. The cache is invalidated when a user is removed, so that the cache always contains the most up-to-date information.

Flag: android.multiuser.cache_user_serial_number
Bug: 340018451
Test: atest com.android.server.pm.UserManagerTest
Change-Id: I1ac96c5777b617803a731109005f75fdb554db98