Remove Dream/background flicker from Communal->Edit transition.
Note that the resulting transition ends up being a little different with
fingerprint from what it looks like without bouncer or with any of the
"screen-based" bouncers. This is because of how the UI fingerprint
looks, and I couldn't find the root cause. Either way, I think it looks
pretty good (see vids in the bug).

Bug: 354701703
Flag: com.android.systemui.communal_hub
Test: atest CommunalSceneStartableTest EditWidgetsActivityStarterTest
Change-Id: If64a76e2438794ab7d04394eccb3667d8767f91c